backend/requirements.txt
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.34.0
python-dotenv>=1.1.0
pydantic-settings>=2.9.0
google-genai>=1.14.0
python-docx>=1.1.0
